feat: add payer-verification oracle interface, mock oracle, and fix clippy warnings
Resolves #91 — oracle interface (oracle_interface.rs, docs/oracle-design.md)
Resolves #94 — mock oracle contract and 9 integration tests
Resolves #95 — oracle integration guide (docs/oracle-integration.md)
Resolves #97 — clippy CI job + zero-warnings across all crates

Closes #91, closes #94, closes #95, closes #97

+++ contracts/iln_governance/src/lib.rs
@@ -422,7 +422,7 @@ impl GovContract {
         if let Some(old_delegate) = Self::get_delegate_raw(&env, &delegator) {
             let old_terminal = Self::resolve_terminal(&env, &old_delegate);
             let delegator_balance = Self::get_own_balance_for_delegation(&env, &delegator);
-            Self::adjust_delegated_to_me(&env, &old_terminal, -(delegator_balance as i128));
+            Self::adjust_delegated_to_me(&env, &old_terminal, -delegator_balance);
         }
 
         // ── Store forward pointer ─────────────────────────────────
@@ -432,7 +432,7 @@ impl GovContract {
 
         // ── Add weight to new terminal ────────────────────────────
         let delegator_balance = Self::get_own_balance_for_delegation(&env, &delegator);
-        Self::adjust_delegated_to_me(&env, &terminal, delegator_balance as i128);
+        Self::adjust_delegated_to_me(&env, &terminal, delegator_balance);
 
         env.events().publish_event(&VotesDelegated {
             delegator,
@@ -453,7 +453,7 @@ impl GovContract {
         if let Some(old_delegate) = Self::get_delegate_raw(&env, &delegator) {
             let old_terminal = Self::resolve_terminal(&env, &old_delegate);
             let delegator_balance = Self::get_own_balance_for_delegation(&env, &delegator);
-            Self::adjust_delegated_to_me(&env, &old_terminal, -(delegator_balance as i128));
+            Self::adjust_delegated_to_me(&env, &old_terminal, -delegator_balance);
 
             env.storage()
                 .persistent()
